#41 innerclass제거 - getOtherConAdapter와 data 분리
